src/requirements.txt
#  pip install -r requirements.txt 
# Set-ExecutionPolicy -ExecutionPolicy RemoteSigned -Scope Process
# Set-ExecutionPolicy Unrestricted -Scope Process 
# python.exe -m pip install --upgrade pip

# & "D:\Coding\PyEnv\.venv\Scripts\Activate.ps1"

# python -c "import sys; print(sys.version)"
# python -c "import sys; print(sys.executable)"
# C:\Users\ASUS\AppData\Local\Programs\Python\Python312\python.exe -m venv --upgrade D:\Coding\PyEnv\.venv

# pip install transformers datasets peft accelerate bitsandbytes torch

# python -m spacy download en_core_web_sm

flask
flask-sqlalchemy
flask_session
mysql-connector-python
torch
torchvision
transformers
# tensorflow
keras
cython
wheel
scikit-learn
python-dotenv
langchain
# hnswlib
# chroma-hnswlib
datasets
huggingface_hub
selenium
webdriver_manager
requests
beautifulsoup4
sentencepiece
SQLAlchemy
accelerate>=0.26.0
psycopg2-binary 
flask-cors
bcrypt
peft 
bitsandbytes
ollama
sentence_transformers
llama_cpp


# Optional (Uncomment if needed)
# numpy==1.23.5
# --upgrade pip setuptools wheel
# spacy
# crewai
# hnswlib
# # pip install spacy[transformers]
# # pip install cython blis murmurhash



# pip list
# deactivate  # Exit if the venv is active
# Remove-Item -Recurse -Force D:\Coding\PyEnv\.venv  # Delete old venv
# C:\Users\ASUS\AppData\Local\Programs\Python\Python312\python.exe -m venv D:\Coding\PyEnv\.venv  # update the path
# D:\Coding\PyEnv\.venv\Scripts\Activate    # activate new venv
# python --version



# code line I run to activate environment are 
# conda create -p venv python==3.10
# conda activate venv
# pip install -r requirements.txt
# python test.py
# pip install ipykernel
# conda install -p d:\Coding\Python_environment\venv ipykernel --update-deps -- force-reinstall
# curl -o python-3.10.0.exe https://www.python.org/ftp/python/3.10.0/python-3.10.0.exe
# python --version